+++ Overlapp.Shared/Model/Api/SearchMovieResponse.cs
@@ -6,7 +6,7 @@ namespace Overlapp.Shared.Model
 	public enum MediaType { Unknown, Movie, Tv }
 
 	public record SearchMovieResponse(int page, SearchMovieRecord[] results, int total_pages, int total_results);
-	public record SearchMovieRecord(bool adult, string backdrop_path, int[] genre_ids, long id, string original_language, string original_title, string overview, float popularity, string poster_path, string release_date, string title, bool video, float vote_average, int vote_count) : IMediaRecord
+	public record SearchMovieRecord(bool adult, string backdrop_path, int[] genre_ids, int id, string original_language, string original_title, string overview, float popularity, string poster_path, string release_date, string title, bool video, float vote_average, int vote_count) : IMediaRecord
 	{
 		public string NameOrTitle => title;
 		public MediaType MediaType => MediaType.Movie;
@@ -22,6 +22,6 @@ namespace Overlapp.Shared.Model
 		string Image { get; }
 		DateTime ReleaseDate { get; }
 		string MediaOverview { get; }
-		long id { get; }
+		int id { get; }
 	}
 }
